services/monitor/requirements.min.txt
flask
google-api-python-client
python-dotenv
requests